perf(maintainability) :recycle: Maintainability improvments
maintainability improvments on `updateData.js`
